// src/app/models/data.ts
export interface Data {
    cod_tend: number;
    date: string;
    dd: number;
    ff: number;
    hbas: number;
    id: number;
    n: number;
    nbas: number;
    numer_sta: number;
    per: number;
    pmer: number;
    pres: number;
    raf10: number;
    rafper: number;
    rr12: number;
    t: number;
    td: number;
    tend: number;
    tend24: number;
    tminsol: number;
    tn12: number;
    tx12: number;
    u: number;
    vv: number;
    ww: number;
  }
  